Use doubles for intermediate calculations.

+++ mesh/shapes.py
@@ -1,13 +1,13 @@
 import numpy as np
 
-from numpy        import float32
+from numpy        import single, double
 from bpy.types    import Object, Depsgraph
 from numpy.typing import NDArray
 
 from .objects     import evaluate_obj
 
 
-def get_shape_mix(source_obj: Object, extra_key: str="") -> NDArray[float32]:
+def get_shape_mix(source_obj: Object, extra_key: str="") -> NDArray[single]:
     """
     Get current mix of shape keys from a mesh. 
     The extra key is optional and will blend that key in at its full value.
@@ -16,7 +16,7 @@ def get_shape_mix(source_obj: Object, extra_key: str="") -> NDArray[float32]:
     vert_count = len(source_obj.data.vertices)
     co_bases   = {}
 
-    basis_co = np.zeros(vert_count * 3, dtype=np.float32)
+    basis_co = np.zeros(vert_count * 3, dtype=double)
     key_blocks[0].data.foreach_get("co", basis_co)
     co_bases[key_blocks[0].name] = basis_co
 
@@ -27,7 +27,7 @@ def get_shape_mix(source_obj: Object, extra_key: str="") -> NDArray[float32]:
     for key in key_blocks[1:]:
         if key.relative_key.name in co_bases:
             continue
-        relative_co = np.zeros(vert_count * 3, dtype=np.float32)
+        relative_co = np.zeros(vert_count * 3, dtype=double)
         key.relative_key.data.foreach_get("co", relative_co)
         co_bases[key.relative_key.name] = relative_co
 
@@ -41,14 +41,14 @@ def get_shape_mix(source_obj: Object, extra_key: str="") -> NDArray[float32]:
             
         if keep or (key.value != 0 and not key.mute):
             relative_key = key.relative_key.name
-            key_coords   = np.zeros(vert_count * 3, dtype=np.float32)
+            key_coords   = np.zeros(vert_count * 3, dtype=double)
             key.data.foreach_get("co", key_coords)
 
             # Use relative key coordinates to get offset
-            offset = key_coords - co_bases[relative_key]
+            offset      = key_coords - co_bases[relative_key]
             mix_coords += offset * blend
     
-    return mix_coords
+    return mix_coords.astype(single)
 
 def create_co_cache(co_cache: dict[str, None], shapes: dict[str, Object], target: Object, base_key: str, vert_count: int, depsgraph: Depsgraph) -> None:
     '''Takes a dict of shape key names of relative keys on a mesh and creates a cache of their coordinates per key.'''
@@ -56,7 +56,7 @@ def create_co_cache(co_cache: dict[str, None], shapes: dict[str, Object], target
         shape = shapes[key_name]
         evaluate_obj(shape, depsgraph)
 
-        shape_co = np.zeros(vert_count * 3, dtype=np.float32)
+        shape_co = np.zeros(vert_count * 3, dtype=double)
         shape.data.vertices.foreach_get("co", shape_co)
 
         co_cache[key_name] = shape_co
@@ -65,15 +65,15 @@ def create_co_cache(co_cache: dict[str, None], shapes: dict[str, Object], target
             target.data.shape_keys.key_blocks[0].data.foreach_set("co", shape_co)
         else:
             new_shape = target.data.shape_keys.key_blocks.get(key_name)
-            new_shape.data.foreach_set("co", shape_co)
+            new_shape.data.foreach_set("co", shape_co.astype(single))
 
         del shapes[key_name]
 
-def create_shape_keys(co_cache: dict[str, NDArray[float32]], shapes: dict[str, Object], target: Object, base_key:str, vert_count: int, depsgraph: Depsgraph) -> None:
+def create_shape_keys(co_cache: dict[str, NDArray[double]], shapes: dict[str, Object], target: Object, base_key:str, vert_count: int, depsgraph: Depsgraph) -> None:
     '''Uses a co_cache to recreate shape keys on target mesh.'''
     for key_name, shape in shapes.items():
         evaluate_obj(shape, depsgraph)
-        shape_co = np.zeros(vert_count * 3, dtype=np.float32)
+        shape_co = np.zeros(vert_count * 3, dtype=double)
         shape.data.vertices.foreach_get("co", shape_co)
 
         rel_key = target.data.shape_keys.key_blocks.get(key_name).relative_key
@@ -81,7 +81,8 @@ def create_shape_keys(co_cache: dict[str, NDArray[float32]], shapes: dict[str, O
         if rel_key.name not in co_cache:
             continue
  
-        offset = shape_co - co_cache[base_key]
+        offset     = shape_co - co_cache[base_key]
         mix_coords = co_cache[rel_key.name] + offset
-        new_shape.data.foreach_set("co", mix_coords)
+        new_shape.data.foreach_set("co", mix_coords.astype(single))
+
     target.data.update()
